# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nursing App API Documentation",
        "description": "API documentation for the Nursing App - A platform connecting users with nursing services",
        "contact": {
            "name": "API Support",
            "email": "support@nursingapp.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/nurses": {
            "get": {
                "tags": [
                    "Nurses"
                ],
                "summary": "List all nurses",
                "description": "Retrieve a list of all nurses. Available to both users and admins.",
                "operationId": "c87e25004d06600f20391cb33066739b",
                "responses": {
                    "200": {
                        "description": "Nurses list retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nurses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Sarah Johnson"
                                                    },
                                                    "phone_number": {
                                                        "type": "string",
                                                        "example": "+1234567890"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "123 Medical Center Dr"
                                                    },
                                                    "profile_picture": {
                                                        "type": "string",
                                                        "example": "https://example.com/photo.jpg"
                                                    },
                                                    "gender": {
                                                        "type": "string",
                                                        "enum": [
                                                            "male",
                                                            "female"
                                                        ],
                                                        "example": "female"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/nurses": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new nurse (Admin only)",
                "description": "Create a new nurse account. Only accessible by admins.",
                "operationId": "a06bf4b94af38ed640a69572b7086f60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone_number",
                                    "address",
                                    "gender"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nurse's full name",
                                        "type": "string",
                                        "example": "Sarah Johnson"
                                    },
                                    "phone_number": {
                                        "description": "Nurse's phone number",
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "address": {
                                        "description": "Nurse's address",
                                        "type": "string",
                                        "example": "123 Medical Center Dr"
                                    },
                                    "profile_picture": {
                                        "description": "URL to profile picture",
                                        "type": "string",
                                        "example": "https://example.com/photo.jpg"
                                    },
                                    "gender": {
                                        "description": "Nurse's gender",
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "female"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nurse created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nurse added successfully."
                                        },
                                        "nurse": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Sarah Johnson"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "+1234567890"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "123 Medical Center Dr"
                                                },
                                                "profile_picture": {
                                                    "type": "string",
                                                    "example": "https://example.com/photo.jpg"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "female"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin role required"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}